+++ src/runconf_ui/config_files/detector_configs/np02_configuration.yml
