+++ src/runconf_ui/config_files/detector_configs/np04_configuration.yml
