+++ src/runconf_ui/screens/quit_screen.py
@@ -7,8 +7,7 @@ from pathlib import Path
 from runconf_ui.runconf_ui_controllers.runconf_ui_state import (
     ShifterInterfaceState,
 )
-from runconf_ui.file_io.save_file_handler import SaveFileHandler
-
+from runconf_ui.utils.save_file_handler import SaveFileHandler
 
 class QuitScreen(Screen):
